% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTumorNormal.R
\name{getTumorNormal}
\alias{getTumorNormal}
\title{Split The Data Into Two Groups}
\usage{
getTumorNormal(betaMatrix)
}
\arguments{
\item{betaMatrix}{A matrix or data frame with column names each at least 15
characters long, where the 14th and 15th characters are both numeric digit}
}
\value{
A list composed of two groups of data
}
\description{
Split The Data Into Two Groups
}
\examples{
\donttest{da<-matrix(sample(1:30,25),nrow=5,ncol=5)
colnames(da)<-paste0("xxxxxxxxxxxxx",c("01y","12y","03y","11y","05y"))
dat<-getTumorNormal(da)
print(dat)}
}
